model pf

+++ CARS/definitions.h
@@ -36,7 +36,7 @@ using Eigen::VectorXd;
 
 // ---------- Globals ----------
 // Radius of circle (squared) surrounding the car. Used in controller.
-const float gCarRadius = (TRACK_WIDTH / 2 + 0.05)*(TRACK_WIDTH / 2 + 0.05)*0.8;
+const double gCarRadius = (TRACK_WIDTH / 2 + 0.05)*(TRACK_WIDTH / 2 + 0.05)*0.8;
 
 // ---------- Enums ----------
 namespace CarMode
